Изменена архитектура. Сервисы стали интерфейсами, на основе их реализованы классы.
