+++ src/main/java/com/ncedu/fooddelivery/api/v1/controllers/RegistrationController.java
@@ -4,7 +4,6 @@ import com.ncedu.fooddelivery.api.v1.dto.RegistrationDTO;
 import com.ncedu.fooddelivery.api.v1.dto.isCreatedDTO;
 import com.ncedu.fooddelivery.api.v1.services.RegistrationService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
